FIX: Handle BOM in files
Switch to UTF-8-sig to handle the BOM in case it is there in the file. Otherwise the file loader may prepend the first CSV field name with the BOM. That will later lead it to believe the file is corrupted, because one of the required fieldnames is missing. The new encoding fixes that.